Update widget code
