+++ src/heretic/id_func.h
@@ -1,8 +1,9 @@
 //
 // Copyright(C) 1993-1996 Id Software, Inc.
 // Copyright(C) 2005-2014 Simon Howard
-// Copyright(C) 2011-2017 RestlessRodent
-// Copyright(C) 2018-2023 Julia Nechaevskaya
+// Copyright(C) 2014-2017 RestlessRodent
+// Copyright(C) 2015-2018 Fabian Greffrath
+// Copyright(C) 2016-2023 Julia Nechaevskaya
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -24,10 +25,15 @@
 
 
 
+extern char ID_Local_Time[64];
+
+//
+// Render Counters and Widgets
+//
+
+extern void ID_RightWidgets (void);
 
 
-#define NUMPLANEBORDERCOLORS 16
-extern int  CRL_PlaneBorderColors[NUMPLANEBORDERCOLORS];
 
 extern int CRL_counter_tome;
 extern int CRL_counter_ring;
@@ -35,11 +41,9 @@ extern int CRL_counter_shadow;
 extern int CRL_counter_wings;
 extern int CRL_counter_torch;
 
-extern void CRL_DrawFPS (void);
-
 extern void CRL_DrawTargetsHealth (void);
 
 // [crispy] demo progress bar and timer widget
-extern void CRL_DemoTimer (const int time);
-extern void CRL_DemoBar (void);
+extern void ID_DemoTimer (const int time);
+extern void ID_DemoBar (void);
 extern int  defdemotics, deftotaldemotics;
